api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
python-multipart==0.0.18
google-generativeai==0.8.3
PyPDF2==3.0.1
Pillow==10.4.0
aiofiles==24.1.0
python-dotenv==1.1.1
numpy==1.26.4